audit: fix silent snapshot-dropping blindspot (float weight 422) + visibility + RL-state ignore
Audit sweep findings + fixes:

1. CRITICAL blindspot — recurring /v1/ingest/snapshot 422 (every ~2s) silently dropped
   the bot's live snapshot: the bridge sends weight as a float (370.6) but the sidecar's
   Vitals.weight and InventoryDigest.weight were int, so pydantic raised int_from_float
   and rejected the ENTIRE snapshot. The sidecar never saw that bot's frame. Fix:
   weight/weight_max are float in state.py. Verified: 0 validation failures after restart.

2. Observability — app.py request-validation handler now puts path + first_error + body
   in the visible log message (was silent structured-only), so the next 422 is actionable
   instead of a blindspot.

3. Hygiene — .gitignore now excludes the ReinforcementLearner (subconscious) runtime
   learned state (data/reinforcement_*.{db,jsonl,pkl,json}) so it isn't committed; that
   is cross-session learning state, not source.

4. Kept the legit discovered knowledge: tables/portals.txt now records iz_ac01_a<->izlude_a
   (server-agnostic DynamicPortalDiscovery persisting the observed portal — de-hardcoding),
   and tables/item_weights.txt has the academy item weights. Reverted the runaway
   mon_control.txt duplicate-append churn (separate dedup flaw, to fix in the writer).

+++ AI_sidecar/ai_sidecar/contracts/state.py
@@ -22,8 +22,12 @@ class Vitals(BaseModel):
     hp_max: int | None = None
     sp: int | None = None
     sp_max: int | None = None
-    weight: int | None = None
-    weight_max: int | None = None
+    # weight is legitimately fractional in RO (sum of per-item weights from the
+    # bridge), so it must be float — declaring it int caused the bridge's float
+    # weight (e.g. 370.6) to 422 (int_from_float) and drop the ENTIRE snapshot
+    # every poll, silently starving the sidecar of live bot state.
+    weight: float | None = None
+    weight_max: float | None = None
     weight_ratio: float | None = None
     hp_ratio: float | None = None
     sp_ratio: float | None = None
@@ -47,8 +51,10 @@ class InventoryDigest(BaseModel):
 
     zeny: int | None = None
     item_count: int | None = None
-    weight: int | None = None
-    weight_max: int | None = None
+    # weight/weight_max are fractional (sum of per-item weights from the bridge) —
+    # must be float or the bridge's 370.6 422s (int_from_float) and drops the snapshot.
+    weight: float | None = None
+    weight_max: float | None = None
     weight_ratio: float | None = None
